end of timeboxed work
